Adjustments to Wowhead parsing, to fix some errors, and be more resilient when errors do occur.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C30408

+++ Rawr.Base/Wowhead.cs
@@ -394,8 +394,8 @@ namespace Rawr
 					break;
 
 				case "armorbonus":
-					item.Stats.Armor -= int.Parse(value);
-					item.Stats.BonusArmor = int.Parse(value);
+					item.Stats.Armor -= float.Parse(value);
+					item.Stats.BonusArmor = float.Parse(value);
 					break;
 				
 				case "healthrgn":
